refactor(renderer-svg): drop CDN URL builder; rename cdn-icons -> icon-dims
The CDN auto-URL path (vendor-keyed) is gone — catalog entries now
carry explicit URLs, so the renderer is just a consumer. Rename the
file to reflect what's left: dimension resolution and icon byte
fetching for URL embedding.

- File: cdn-icons.ts → icon-dims.ts (git rename)
- Drop: getCDNIconUrl / hasCDNIcons / getIconExtension /
  resolveIconDimensionsForGraph / DEFAULT_CDN_CONFIG / CDNConfig /
  ResolvedIconDimensions / SVG_VENDORS / PNG_VENDORS / normalizeModel
- Rename: fetchCDNIcon → fetchIconAsDataUrl
- Keep: fetchImageDimensions, resolveAllIconDimensions, IconDimensions

calculateIconInfo and the subgraph icon path in svg.ts now go through
core's resolveIcon for the inline/URL split. collectIconUrls returns
only URL-form icons (inline doesn't need fetching).

pipeline.ts surfaces iconDimensions as a flat Map<url, dims> instead
of the byUrl/byKey wrapper — the byKey side was used only by the
deleted CDN flow.
