+++ libs/@shumoku/renderer-svg/src/svg.ts
@@ -29,24 +29,18 @@ import {
   computeNodeSize,
   DEFAULT_ICON_SIZE,
   darkTheme,
-  getDeviceIcon,
   ICON_LABEL_GAP,
   LABEL_LINE_HEIGHT,
   lightTheme,
   linkSpeedBps,
+  resolveIcon,
   SMALL_LABEL_CHAR_WIDTH,
   type SurfaceToken,
   specDeviceType,
-  specIconKey,
   type Theme,
 } from '@shumoku/core'
 
-import {
-  getCDNIconUrl,
-  hasCDNIcons,
-  type IconDimensions,
-  resolveAllIconDimensions,
-} from './cdn-icons.js'
+import { type IconDimensions, resolveAllIconDimensions } from './icon-dims.js'
 import type { DataAttributeOptions, RenderMode } from './types.js'
 
 // ============================================
@@ -668,49 +662,26 @@ export class SVGRenderer {
 
     const rx = 12 // Border radius (larger for container/card feel)
 
-    // Check if subgraph has icon (user-specified or CDN-supported vendor)
-    // AWS uses service/resource format (e.g., ec2/instance)
+    // Subgraphs only render URL icons here (inline SVG drawing inside a
+    // container is uncommon enough to keep this path simple).
     const sgSpec = subgraph.spec
-    const iconKey = specIconKey(sgSpec)
     let hasIcon = false
     const defaultIconSize = 24
     const iconPadding = 8
 
-    // Get icon URL and dimensions for sizing calculations
     let iconUrl: string | undefined
     let iconWidth = defaultIconSize
     let iconHeight = defaultIconSize
 
-    if (sgSpec?.icon) {
+    if (sgSpec?.icon && !sgSpec.icon.trim().startsWith('<')) {
       hasIcon = true
       iconUrl = sgSpec.icon
       const dims = this.options.iconDimensions.get(sgSpec.icon)
       if (dims) {
         const aspectRatio = dims.width / dims.height
-        if (aspectRatio >= 1) {
-          iconHeight = defaultIconSize
-          iconWidth = Math.round(defaultIconSize * aspectRatio)
-        } else {
-          iconHeight = defaultIconSize
-          iconWidth = Math.round(defaultIconSize * aspectRatio)
-        }
+        iconHeight = defaultIconSize
+        iconWidth = Math.round(defaultIconSize * aspectRatio)
       }
-    } else if (sgSpec?.vendor && iconKey && hasCDNIcons(sgSpec.vendor)) {
-      const cdnUrl = getCDNIconUrl(sgSpec.vendor, iconKey)
-      const dims = this.options.iconDimensions.get(cdnUrl)
-      if (dims) {
-        hasIcon = true
-        iconUrl = cdnUrl
-        const aspectRatio = dims.width / dims.height
-        if (aspectRatio >= 1) {
-          iconHeight = defaultIconSize
-          iconWidth = Math.round(defaultIconSize * aspectRatio)
-        } else {
-          iconHeight = defaultIconSize
-          iconWidth = Math.round(defaultIconSize * aspectRatio)
-        }
-      }
-      // dims not found = icon doesn't exist on CDN, skip icon
     }
 
     // Calculate icon position (top-left corner)
@@ -1084,53 +1055,23 @@ ${fg}
     // Use full node width as max; layout engine already calculated appropriate size
     const maxIconWidth = w
 
-    const spec = node.spec
+    const icon = resolveIcon(node.spec)
+    if (!icon) return null
 
-    // User-specified icon takes highest priority. Either inline SVG content
-    // (`<path .../>` or `<svg ...>...</svg>`) or a URL.
-    if (spec?.icon) {
-      if (spec.icon.trim().startsWith('<')) {
-        return {
-          width: DEFAULT_ICON_SIZE,
-          height: DEFAULT_ICON_SIZE,
-          svg: `<svg width="${DEFAULT_ICON_SIZE}" height="${DEFAULT_ICON_SIZE}" viewBox="0 0 24 24" fill="currentColor">${spec.icon}</svg>`,
-        }
-      }
-      const dims = this.options.iconDimensions.get(spec.icon)
-      const { width, height } = this.calculateIconSize(dims, maxIconWidth)
+    if (icon.kind === 'inline') {
       return {
-        width,
-        height,
-        svg: `<image href="${spec.icon}" width="${width}" height="${height}" preserveAspectRatio="xMidYMid meet" />`,
+        width: DEFAULT_ICON_SIZE,
+        height: DEFAULT_ICON_SIZE,
+        svg: `<svg width="${DEFAULT_ICON_SIZE}" height="${DEFAULT_ICON_SIZE}" viewBox="0 0 24 24" fill="currentColor">${icon.svg}</svg>`,
       }
     }
 
-    // Try vendor-specific icon first (service for cloud, model for hardware)
-    // AWS uses service/resource format (e.g., ec2/instance)
-    const iconKey = specIconKey(spec)
-    // Use CDN icons for supported vendors (only if dimensions were resolved, i.e. icon exists)
-    if (spec?.vendor && iconKey && hasCDNIcons(spec.vendor)) {
-      const cdnUrl = getCDNIconUrl(spec.vendor, iconKey)
-      const dims = this.options.iconDimensions.get(cdnUrl)
-      if (dims) {
-        const { width, height } = this.calculateIconSize(dims, maxIconWidth)
-        return {
-          width,
-          height,
-          svg: `<image href="${cdnUrl}" width="${width}" height="${height}" preserveAspectRatio="xMidYMid meet" />`,
-        }
-      }
-      // dims not found = icon doesn't exist on CDN, fall through to device type icon
-    }
-
-    // Fall back to device type icon
-    const iconPath = getDeviceIcon(specDeviceType(spec))
-    if (!iconPath) return null
-
+    const dims = this.options.iconDimensions.get(icon.url)
+    const { width, height } = this.calculateIconSize(dims, maxIconWidth)
     return {
-      width: DEFAULT_ICON_SIZE,
-      height: DEFAULT_ICON_SIZE,
-      svg: `<svg width="${DEFAULT_ICON_SIZE}" height="${DEFAULT_ICON_SIZE}" viewBox="0 0 24 24" fill="currentColor">${iconPath}</svg>`,
+      width,
+      height,
+      svg: `<image href="${icon.url}" width="${width}" height="${height}" preserveAspectRatio="xMidYMid meet" />`,
     }
   }
 
@@ -1849,28 +1790,21 @@ export function render(graph: NetworkGraph, layout: LayoutResult, options?: Rend
 }
 
 /**
- * Collect all icon URLs from a NetworkGraph
+ * Collect every URL-form icon referenced by the graph, so a caller can
+ * pre-resolve their dimensions (or fetch their bytes) before rendering.
+ * Inline SVG icons (`<...>`) are skipped — they don't need fetching.
  */
 export function collectIconUrls(graph: NetworkGraph): string[] {
   const urls = new Set<string>()
+  const isUrl = (icon: string | undefined): icon is string => !!icon && !icon.trim().startsWith('<')
 
   for (const node of graph.nodes) {
-    const spec = node.spec
-    if (spec?.icon) {
-      urls.add(spec.icon)
-    } else if (spec?.vendor && hasCDNIcons(spec.vendor)) {
-      const iconKey = specIconKey(spec)
-      if (iconKey) {
-        urls.add(getCDNIconUrl(spec.vendor, iconKey))
-      }
-    }
+    if (isUrl(node.spec?.icon)) urls.add(node.spec.icon as string)
   }
 
   if (graph.subgraphs) {
     for (const sg of graph.subgraphs) {
-      if (sg.spec?.icon) {
-        urls.add(sg.spec.icon)
-      }
+      if (isUrl(sg.spec?.icon)) urls.add(sg.spec.icon as string)
     }
   }
